refactor: 1차 리팩토링
- 크롤링, 알림, 스토리지 파일 구분
- img 파일 경로 이전
