+++ .eslintrc.json
@@ -1,7 +1,8 @@
 {
   "env": {
     "browser": true,
-    "es2021": true
+    "es2021": true,
+    "webextensions": true
   },
   "extends": ["airbnb-base", "eslint:recommended", "plugin:prettier/recommended"],
   "parserOptions": {
